+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/dialog/filter-by-column/o-table-filter-by-column-data-dialog.component.ts
@@ -227,6 +227,9 @@ export class OTableFilterByColumnDataDialogComponent implements AfterViewInit {
       } else {
         colRenderedValues.forEach((renderedValue, i) => {
           /*Selection distint values */
+          if (renderedValue === null || renderedValue === undefined) {
+            renderedValue = '';
+          }
           if (!this.columnData.find(item => item.renderedValue === renderedValue)) {
             this.addIntoColumnData(renderedValue, colValues, i, filter);
           }