+++ projects/ontimize-web-ngx/src/lib/types/error-form-validation.type.ts
@@ -1,0 +1,5 @@
+export type OFormValidation = {
+  valid: boolean;
+  title?: string;
+  messages?: Array<string>;
+};